add login , registeration and full resources for the 3 stakeholders

+++ app/Http/Controllers/ManagerController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\DataTables\ManagerDatatable;
 use App\Http\Requests\ManagerRequest;
+use App\Http\Requests\StoreManagerRequest;
+use App\Http\Requests\UpdateManagerRequest;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
@@ -39,16 +41,13 @@ class ManagerController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(ManagerRequest $request)
+    public function store(StoreManagerRequest $request)
     {
-
-        $data = array();
         
         $data = $request->all();
 
         if($request->hasFile('avatar_image')){
 
-            // $data['avatar_image'] = $request->avatar_image->getClientOriginalName();
             $data['avatar_image'] = rand() . '.' . $request->avatar_image->getClientOriginalExtension();
             $request->avatar_image->storeAs('users_images', $data['avatar_image']);
 
@@ -103,28 +102,22 @@ class ManagerController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, User $user)
+    public function update(UpdateManagerRequest $request, User $user)
     {
 
-        dd($request->all());
+        $data = $request->all();
 
-        // $data = $request->validate([
+        //delete old image of post if exist and not the default image
+        if($request->hasFile('avatar_image')){
 
-        //     'name'     => 'required',
-        //     'email'    => 'required|email|unique:users,email,'. $user->id,
-        //     'password' => 'sometimes|nullable|min:8',
-        // ]);
+            if (!empty($user->avatar_image) && $user->avatar_image != 'avatar.png') {
+                Storage::delete('users_images/' . $user->avatar_image);
+            }
 
-
-        //delete old image of post if exist
-        if(!empty($user->avatar_image) && $user->avatar_image != 'avatar.png'){
-            Storage::delete('users_images/' . $user->avatar_image );
-        }
-
-        if(!empty($user->avatar_image)){
             $data['avatar_image'] = rand() . '.' . $request->avatar_image->getClientOriginalExtension();
-            $user->avatar_image->storeAs('users_images', $data['post_img']);
-        }
+            $request->avatar_image->storeAs('users_images', $data['avatar_image']);
+
+        } 
 
         if ($request->filled('password')) {
             $data['password'] = Hash::make($data['password']);
